win32: make realpath() return an error if necessary
resolve_symlinks() uses GetFinalPathNameByHandleA which is loaded
at runtime because it isn't available in all versions of Windows.

If GetFinalPathNameByHandleA isn't available resolve_symlinks()
(and hence realpath()) should return a NULL pointer to indicate an
error, not the original path.  Not returning NULL causes an infinite
loop in do_lstat().

See GitHub issue #204.