tls: avoid unnecessary changes to POSIX build
Calls to tls_error_die() embed the line number of the error in the
binary.  Since some lines had been added to tls.c for the WIN32
port a POSIX build of the busybox-w32 source differed from upstream.

Avoid this by pushing the special handling of /dev/urandom down into
open_read_close().  tls.c is now unchanged from upstream.

The only differences in the POSIX build reported by 'objdump -s'
are now the GNU build id, the copyright date and the date of the
configuration.