tls: avoid unnecessary changes to POSIX build, part 2
On reflection, the previous commit may have been ill-advised.  There
are many calls to open_read_close() and most shouldn't be able to
access special devices.  (Though in practice only a few are enabled
in busybox-w32.)

Nonetheless, I've implemented a new mechanism which uses the macro
MINGW_SPECIAL() to mark calls to functions that are allowed to
access special devices.

An unrelated change is to avoid compiling fputs_stdout() in
coreutils/printf.c for the POSIX build.