printf: better support for escape sequences
Upstream printf outputs one character at a time which doesn't
play well with emulation of ANSI escape sequences.  Previous
workarounds for this only applied in limited circumstances.

Try a different approach:  replace putchar() and printf() calls
in the printf applet with custom routines which store the output
in memory.  It's only really output at the end of the program or
when a newline is detected and a non-trivial amount has been
collected.