free: implement sysinfo(2) and enable free(1)
This is an experimental implementation of sysinfo(2)/free(1).

It uses the WIN32 API GlobalMemoryStatusEx() to obtain information
about memory.

It seems that the 'total pagefile' value includes total RAM as well
as pagefile and 'available pagefile' includes available RAM.  So the
RAM values are deducted.

I've no idea what corresponds to Linux buffers and cache.