ash: correctly identify applet in getopt() error messages
With the default configuration most applets are run from the shell
via run_noexec_applet_and_exit().  If the applet runs getopt() and
an error is detected the message takes the form:

   sh: unknown option -- q

This happens because mingw-w64's implementation of getopt() uses
the global variable __argv[0] as the name of the program.

Avoid the misleading message by setting __argv[0] as appropriate.