winansi: fix ansi emulation
The following commands (reported in GitHub issue #201):

   printf "\033[38;2;255;0;0mX\033[m\n"
   printf "\033[38;2;255;0;0m;\033[m\n"

produce different results.  The first correctly displays a red 'X'
while the second incorrectly displays a white ';'.

The problem is that process_24bit() overruns the extent of the
escape sequence.  As a result the loop in process_escape() which
handles 'ESC[...m' sequences sees the ';' in the text as a
continuation of the escape sequence.

Fix this by:

- reworking process_24bit() so that the overrun is avoided;

- changing the test in the loop in process_escape() so that even
  if an overrun happens it stops processing at the end of the escape
  sequence.

Also, save a few bytes by replacing '++str' with 'str + 1' in a
few places.