libbb: fix another tab-completion corner case
Commit 90f35327c (libbb: avoid problems with quoted characters in
tab completion) avoided a conflict between the code to quote
special characters and that to display case-preserved filenames.
Here's another one.

I'd installed a third-party binary on Microsoft Windows and wanted
to execute it from a busybox-w32 shell which didn't have it in its
path.  No problem:

   $ c:/prog<tab>
   $ c:/Program<tab>
   Program Files/  ProgramData
   $ c:/Program\<tab>
   $ c:/PProgram

I was too quick to enter the final <tab>, I'd meant to add a space.

A trailing backslash is effectively removed by the tab completion
matching code but it remains on the display.  This confuses the
code to display case-preserved filenames.

Upstream BusyBox just ends up with a spurious backslash in the name.
Bash rewrites the command line to remove the surplus backslash.

Fix the issue in busybox-w32 by leaving the backslash in the name,
even for matching purposes.  This will never result in a match,
thus affording the user a change to correct their mistake.