mingw: don't include unwind tables in 64-bit build
It seems that mingw-w64 no longer requires unwind tables in the
64-bit build, resulting in a significant reduction in the size of
the binary.

The flags for unwind tables were added in commit 535ec990c (mingw:
include unwind tables in 64-bit build by default).  Removing them
means that the only difference between the 32- and 64-bit config
files is the compiler prefix.