win32: changes to mntent implementation
Make the structure returned by getmntent(3) static so it persists
after endmntent(3) closes the stream.  No current caller in the
WIN32 port needs this functionality but it's good to match the
documented behaviour.

Populate more fields of the mntent structure in find_mount_point().
This is required to support the df -t and -T flags recently added
upstream.

The static structures used here are allocated on demand.  Separate
static structures are used in each case because df iterates through
mounts calling statfs(2) on each and the WIN32 implementation of
statfs(2) calls find_mount_point().