win32: special case for devices files in stat(2)
The diff applet calls stat(2) on the files it's asked to process.
This includes /dev/null when the -N flag is used and /dev/fd/*
files when process substitution is used.

Treat device files as a special case in get_file_attr(), returning
a fake set of attributes with FILE_ATTRIBUTE_DEVICE set.  This
value is unused elsewhere in busybox-w32.  Ensure it's unset in
other cases.

When FILE_ATTRIBUTE_DEVICE is set:

- adjust some permissions;

- avoid calling has_exec_format() as this opens/closes the file
  which breaks process substitution.

These changes improve the behaviour of diff but they also have other
effects.  For example, the stat and ls applets now report details of
device files.  There may be unintended consequences.