ash: reset ANSI emulation at prompt
It appears that MSYS2/Cygwin applications turn off virtual terminal
processing.  This caused literal ANSI escape sequences to be echoed
in the shell, which assumed the console state it had applied would
remain unchanged.

Add a call to skip_ansi_emulation(TRUE) before issuing a shell
prompt to ensure that the state is correct.

GitHub issue #223.