win32: better detection of Alt key release
Commit 7874ca73b (win32: allow characters to be entered as ALTNNN)
made it possible to enter characters using Alt and a character code
on the numeric pad.  This required detecting the release of the
Alt key.

Sometimes this caused a problem when using Alt+Tab to cycle between
applications.  If Alt was released before Tab the code passed a tab
character to the application.

Avoid such issues by looking specifically for the release of the Alt
key.