win32: fix fnmatch(3) handling of xdigit
The glob pattern '[[:xdigit::]]*' didn't return the matches
expected.  It turns out the implementation (from glibc) fails
to detect 'xdigit' as a valid character class.  Changing the
definition of CHAR_CLASS_MAX_LENGTH to 7 fixes the problem.

This was never an issue in glibc because it uses a different
definition.  More modern versions of fnmatch(3) in glibc and
gnulib also make CHAR_CLASS_MAX_LENGTH long enough.

The code for fnmatch(3) was taken from glibc at commit
7814856974388a856a575fa45f88d502c8a1ab29.  This was the last
version before the code was rearraged to better support multibyte
characters.