win32: tidy up fetching of system directory
Add a new function, getsysdir(), to fetch and cache the system
directory.  This avoids the non-intuitive use of getpwuid() in
get_system_drive().

The call to GetSystemDirectory() in get_proc_addr() can't be
replaced because getsysdir() calls realpath() which requires a
call to get_proc_addr().

No change in the size of the binary.