printf: improved error handling
The printf applet is NOFORK and is used to implement the printf
command in ash.  It should therefore be careful about handling
memory allocation failures.

If vasprintf() or realloc() fail the applet now carries on as best
it can.  This may result in some lost output but at least the shell
will survive.

Saves 32 bytes.