vi: add fileformats option
Add a limited version of the 'fileformats' option.  Only the
'dos,unix' and 'unix,dos' settings are allowed with the former
as default.

When a file is read for editing its format is detected.  If it
only has LF line endings it's treated as being of type 'unix'
otherwise its type is 'dos'.

When an empty edit buffer is created it's given the first file
type listed in 'fileformats'.

The file type is stored in the read-only option 'fileformat'
and is used to select the appropriate line endings when the
file is written.

See GitHub issue #220.  Adds 168 bytes to the binary.