ash: restrict valid job ids
Although job control is disabled in busybox-w32 getjob() is still
compiled in because it's used by waitcmd().  Job ids of the form
'%string' or '%?string' don't work, though, because the command
isn't stored in the job structure.  Disable the code to handle
them.

Job ids '%%', '%+', '%-' and '%num' are still allowed and some of
them are (successfully) used in the ash tests.

Saves 144 bytes.