ash: avoid problems with non-zero DEBUG
When DEBUG had a non-zero value the preprocessor complained that
'close' was being redefined.  In a WIN32 build it's set to
'mingw_close'.  Since this isn't necessary in ash it can be
undefined without ill effects.

DEBUG can be set to 1 or 2.  In a WIN32 build this value was being
used to calculate indices for optlist[], giving incorrect results.
Using (DEBUG != 0) instead fixes this.