win32: limited version of timegm(3)
timegm(3) from musl checks that the calculated time_t value can
be broken out into a struct tm without overflow.  The limiting
factor is that tm_year is an int.

Our only use of timegm(3) is followed by a call to localtime(3).
The Microsoft version of this only works for time_t values between
0 and INT_MAX (32-bit) or 0 and 32535215999 (64-bit).

Enforce these limits in timegm(3), thus avoiding the use of musl's
__secs_to_tm() and saving 624 bytes.