ash: prevent exit if background jobs are present
If the user exits from an interactive shell while background jobs
are running the console window won't terminate because the
background jobs are still connected to it.

Warn the user when this happens.  They can either shut down the
background jobs themselves or close the console window.

This is implemented by repurposing the stoppedjobs() function
which is otherwise unused in the WIN32 port.