win32: rename mingw_spawn_1
The name of the function mingw_spawn_1() wasn't particularly
meaningful.  Remove its envp argument (as all of its current callers
pass a NULL pointer) and rename it mingw_spawnvp() to better reflect
what it does.

The path search it performs isn't the standard one:  it has features
specific to busybox-w32.