cpio: code shrink with !FEATURE_EXTRA_FILE_DATA
The new --renumber-inodes option won't work if FEATURE_EXTRA_FILE_DATA
is disabled.  Don't allow it to be selected in that case.  In fact,
if FEATURE_EXTRA_FILE_DATA is disabled all the code to handle hardlinks
can be #ifdef'ed out.

The new --ignore-devno option need do nothing if FEATURE_EXTRA_FILE_DATA
is disabled, as in that case the device ids will already be zero.
It can still be selected, cpio will just always behave as though the
option was provided on the command line.

This doesn't affect the size of the default build.