ash: workaround for UCRT bug
There seems to be a bug in UCRT such that if a process has been
started by passing a non-NULL environment block to CreateProcess()
any subsequent call to spawnve() with a non-NULL environment pointer
results in a crash.

Commit 5b48ca53b (win32: pass NULL to spawnve, not environ) fixed
the problem in busybox-w32 for those cases where a NULL environment
pointer was sufficient.  It didn't handle the case where the shell
passes a modified environment to its child.

All calls to spawnve() in the shell occur in a process which will
terminate whether or not the call succeeds.  It therefore doesn't
matter if we mess with the environment of this process to allow
spawnve() to be passed a NULL environment pointer.  Do this for
UCRT builds only.

(GitHub issue #234)