win32: fake file ownership on FAT filesystems
busybox-w32 tries to display the ownership of files by comparing
the security identifier (SID) of the current user with that of the
file on disk.  For filesystems that don't support SIDs this resulted
in files being listed as owned by root.

It appears that filesystems without support for file ownership
return a null SID.  In such cases pretend the file belongs to the
current user.

GitHub issue #241.