ash: let $HOME set home directory of login shell
In busybox-w32 the shell option '-l' sets HOME to the user's
home directory, as determined by a call to GetUserProfileDirectory().
This is differs from how shells work on Unix, where HOME is usually
set by login(1).

Allow the user to override this behaviour by setting HOME before
starting the shell.  If HOME isn't set or contains an empty string
the previous behaviour applies.  If HOME is set to a non-empty string
the user should ensure that it represents a valid absolute path.

(GitHub issue #244)