ash: avoid misleading '.exe' in standalone shell mode
In dash a command with an explicit path has an index value of -1
in its cmdentry structure.  Upstream BusyBox extends this to
allow other negative values for applets in standalone shell mode.
busybox-w32 adds a further extension, with the index value INT_MIN
indicating an applet with a Unix-style path.

In describe_command() the command should only have a '.exe'
extension added for index == -1, the original dash case of an
explicit path.

(GitHub issue #248)