which,ash: changes to which/command/type
Change how 'which' detects if it was run from a standalone shell:
the shell passes the undocumented '-s' option.  This is stricter
and more reliable than the previous method of checking the name
of the binary.

Add a function to determine the binary associated with a given
applet name.  This makes it possible for 'which' and 'command -v'
to list the correct binary even for applets other than 'busybox'.
For example, when the binary is called 'sh.exe' 'which sh' will
report its path.

In standalone shell mode 'command -V' and 'type' now report "xxx
is a builtin applet" rather than "xxx is xxx", which is true but
not very illuminating.

(GitHub issue #248)