win32: new code to set empty environment variable
Windows environment variables: a never-ending source of fun.

It seems the recent hack to work around problems in UCRT breaks
the hack to set empty environment variables.  The change to the
empty variable isn't reflected in the environment seen by the C
runtime.

Use the WIN32 API to set the empty variable then set a dummy
variable to make the C runtime take notice.

(GitHub issue #250)