win32: allow preference for applets to be disabled at runtime
The default busybox-w32 configuration enables the PREFER_APPLETS
and SH_STANDALONE features.  Sometimes it may be desirable to
override the default preference for applets, for example, if an
applet needs to be replaced by an external program with additional
features.

Add support for the environment variable BB_OVERRIDE_APPLETS.
Its value may be:

- a single dash ('-'):  all applets are overridden;

- a space-separated list of names:  only the specified applets
  are overridden.