ash: export certain variables to the environment immediately
The environment variables BB_OVERRIDE_APPLETS, BB_SKIP_ANSI_EMULATION
and BB_SYSTEMROOT affect of the behaviour of the shell itself.
Setting them as shell variables is insufficient for them to affect
the current shell.

When these three variables are exported from the shell they are
now placed in the environment immediately.  Conversely, when
they're unset or unexported they're removed from the environment.