win32: try to get link count for directories
On Unix the link count of a directory reflects the number of
subdirectories it contains.  Enhance readdir(3) to return file
types and use this to count subdirectories when stat(2) is called
for a directory.

As with other features that might slow down stat(2) this is
controlled by the build-time setting FEATURE_EXTRA_FILE_DATA.
(Commit d82db8e9a 'win32: make stat(2) fetch additional metadata').

(GitHub issue #254)