libarchive: alter handling of failure to create symlinks
Unlike GNU tar upstream BusyBox tar treats all errors as fatal.

Failure to create a symlink on Windows is very likely due to a
lack of the required privilege.  busybox-w32 therefore treats this
case as a non-fatal error and, as far as possible, continues
extracting files.

However:

- The inconsistency of this behaviour caused some confusion.

- busybox-w32 tar didn't return an error status when a non-fatal
  error was encountered, unlike GNU tar.

Change the implementation so extraction continues when symlink
creation fails but the error report is deferred until all of the
archive has been processed.  Only one error is reported and it's
treated as fatal.

Saves 48-64 bytes.

(GitHub issue #265)