build system: fix breakage affecting MSYS2 build
Upstream changes to the embedded_scripts and mkconfigs scripts
resulted in busybox-w32 failing to build with MSYS2.

bzip2 in MSYS2 detects stdout redirected to /dev/null as a terminal
and returns a non-zero exit status.  Since the test is only for
the existence of bzip2, not its functionality, even an exit status
of 1 is OK.  Only fail if the exit status is 127.