win32: allow 'make menuconfig' in native builds
'make menuconfig' is a build option to allow changes to be made
to the BusyBox configuration.  It's a bit friendlier than using
a text editor on the .config file.

Previously 'menuconfig' was only available when cross-compiling
on a Linux/Unix platform.  The 'mconf' build program has been
ported to WIN32 so it's now able to run the 'lxdialog' program
which displays the configuration dialogs.

Building 'lxdialog' is somewhat awkward in MSYS2.  The MINGW32/64
build environments generate WIN32 executables for 'lxdialog'.
These doesn't work properly in the MSYS2 console.

For 'menuconfig' to work it's necessary to run it in an MSYS
build environment.  This generates an MSYS binary which works
in the MSYS2 console.  busybox-w32 should then be built in a
MINGW32/64 build environment.  Doing so will generate additional
copies of the build programs without a '.exe' suffix:  the MSYS
build environment adds '.exe' to binaries it builds.  This breaks
'menuconfig'.

To configure an MSYS build environment use:

   pacman -S gcc ncurses-devel