build system: try harder to detect w64devkit
We need to build the supplied PDCurses code when using w64devkit.
This was being detected by checking for the W64DEVKIT environment
variable, but this is only defined if w64devkit is started via
w64devkit.exe.

Set W64DEVKIT ourselves if HOSTCC targets the mingw32 platform.
This won't be the case when cross-compiling on Linux but will
for w64devkit and MSYS2 MINGW32/64.

The build won't work properly for MSYS2 MINGW32/64, but it doesn't
work when using the supplied curses library either.  'make menuconfig'
requires the use of MSYS2 MSYS, and HOSTCC there targets msys.