make: changes to suffix substitution in macro expansion
The POSIX standard defines suffix substitution in macro expansion
as taking the form:

   $(string1 [: subst1 =[ subst2 ]])

Since 'subst1' isn't bracketed a value must be supplied.  Enforce
this in POSIX mode.

As a non-POSIX extension an empty 'subst1' is permitted with 'subst2'
being added to all words unconditionally.  If both 'subst1' and
'subst2' are empty the words are returned unchanged.