getopt: use name supplied with '-n' to report error
On Windows:

   $ getopt -o a -n my_script -- -z
   getopt: unknown option -- z

On Linux:

   $ getopt -o a -n my_script -- -z
   my_script: unknown option -- z

The difference arises because the Mingw-w64 implementation of
getopt(3) uses __argv[0] to report errors, not the first element
of the argument array passed to it.

Make __argv[0] point to the name supplied with the '-n' option to
match behaviour when glibc getopt(3) is used.

(GitHub issue #274)