win32: prevent stat(2) from updating access times
The WIN32 implementation of stat(2) reads the contents of some files
to see if they're executable.  This may update the file access time.
Avoid this by a special call to SetFileTime() after opening the file.

For details see:

   https://devblogs.microsoft.com/oldnewthing/20111010-00/?p=9433
   https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-setfiletime

File access times aren't updated by default in recent versions of
Windows.  This commit is only necessary if updating of file access
times is explicitly enabled:

   fsutil behavior set DisableLastAccess 0