awk: make random values more random
When srand(3) is called to seed the random number generator with
the current time the first value returned by rand(3) changes slowly
with time.  This is a property of the implementation in the C runtime.

Change the order in which values from rand(3) are consumed to generate
the value returned by the awk rand() function.  This puts the value
returned by the first call to rand(3) in the least significant bits,
not the most significant.

(GitHub issue #279)