awk: further improvements to randomness
POSIX requires that the awk srand() function uses the time of day
to seed the PRNG.  The obvious implementation used in BusyBox does
exactly that, passing time(NULL) to srand(3).

When processes are started within a few seconds of one another their
seeds are very similar.  Given the realtively poor quality of rand(3)
in some C runtimes this results in random number sequences that are
somewhat correlated.

Improve matters by using an integer hash on the seed, as recommended
here:

   https://nullprogram.com/blog/2019/04/30/#the-wrong-places

Costs 48 bytes.

(GitHub issue #279)