win32: improve seeding of PRNGs
busybox-w32 provides two PRNG implementations which are used in
the emulation of /dev/urandom.  The ad hoc method of seeding them
has been replaced by calls to RtlGenRandom.

The documentation for RtlGenRandom indicates it has been deprecated
in favour of CryptGenRandom.  The documentation for the latter
indicates it has been deprecated in favour of the 'Cryptography Next
Generation APIs'.  Nonetheless, RtlGenRandom remains available in
every version of Windows since XP.

In the unlikely event that RtlGenRandom fails simply use the current
time as the seed.

Saves 192 bytes in the default configuration.