win32: use ACL check to clarify write permission
On Microsoft Windows a user's home directory doesn't belong to
them:  it actually belongs to the 'system' user.  stat(2) was
only using ownership to determine write permissions, so it seemed
that the user was unable to write to their own home directory.

Use a call to AccessCheck() to determine if files can be accessed
due to an entry in their ACL.  User home directories and a few other
files (e.g. C:/Users/Public) now have the correct write permission.

This feature is enabled by FEATURE_EXTRA_FILE_DATA.

Costs 220-256 bytes.

(GitHub issue #280)