win32: minor improvements to stat(2)
Commit b11352dcb (win32: prevent stat(2) from updating access times)
requested GENERIC_ALL access when opening files.  It appears that
GENERIC_READ is sufficient and also faster.

The code to find the actual size of compressed or sparse files only
needs to be invoked for regular files.  Avoiding unnecessary calls
to GetCompressedFileSize() makes stat(2) slightly faster and gives
a more accurate number of blocks for symbolic links.