win32: more minor improvements to stat(2)
The previous commit incorrectly stated that preventing the access
time of a file from being updated only required it to be opened
with GENERIC_READ access.  In fact, even though we don't want to
update the access time, SetFileTime() also requires the file to
have been opened with FILE_WRITE_ATTRIBUTES access.

There's no need to explicitly avoid device files when checking
for execute mode:  since device files are now 'character special'
they are excluded by the test that the file is 'regular'.

Device files should be excluded when trying to obtain extra file
data using GetFileInformationByHandle().  It shouldn't be possible
for CreateFile() to open then, so there's no point in trying.