win32: limit setting of errno when lazy loading fails
The function get_proc_addr() facilitates dynamic loading of functions
from DLLs.  In the event of failure it set errno to ENOSYS.

This is sometimes useful, such as in the implementations of link(2),
symlink(2) and realpath(3).  However, many other uses of lazy loading
can recover from failure or simply don't care.  In these cases setting
errno is unnecessary and may be counterproductive.

(GitHub issue #283)