diff: implement --binary flag
On Windows GNU diff uses text mode for input and output.  It also
has the '--binary' flag to use binary mode instead.  On Unix binary
mode is the default and the flag does nothing.

Alter diff to use text mode by default for input (though not output,
let's not go overboard).  Add the '--binary' flag to override this.

Costs 96-160 bytes.