win32: provide a default value for HOME
The busybox-w32 shell initialises HOME when it starts.  However,
if applets are run outside the environment provided by the shell
they may find HOME is unset.  This caused a problem for 'vi' as
it was unable to locate its .exrc.

If HOME isn't available in the environment make getenv(3) provide
a sensible default value.  The shell must use the *real* getenv(3)
when determining if HOME is already set.

Also, unrelated to the above, the shell shouldn't treat failure of
getpwuid(3) as a fatal error.

Costs 72-80 bytes.