libbb: fix CRLF handling
Ensure a trailing CR is only removed if it precedes a LF.

The two cases at issue are intended to read complete lines and remove
the line terminator.  In the normal case a trailing LF will be present
so removing the CR unconditionally worked.  However, if the last line
of a file was missing its LF or if a NUL was detected (in the case of
xmalloc_fgetline()) the CR might have been removed without justification.