make: strip leading whitespace in shell assignment
Assignment of shell output to a macro ('!=') was originally a
non-POSIX extension.  It later became a POSIX 202X feature.

However, the implementation failed to include the additional
POSIX requirement that leading whitespace is removed from the
shell output.

Neither GNU make nor bmake strip leading whitespace.  Implement
this behaviour as a non-POSIX extension.