win32: work around problem with ConEmu
When running the busybox-w32 shell in ConEmu, if:

- ANSI emulation is disabled (which it will be, by default)
- a non-builtin command is run

there is a window of a few seconds after the command completes during
which ConEmu is in XTerm mode.  During this time any attempt to use
the arrows keys will result in [A[B[C[D appearing.  This appears to
be a common problem:

   https://github.com/Maximus5/ConEmu/issues/2316

Try to detect if we're running under ConEmu and alter the default
behaviour to prefer ANSI emulation.  The user can override this
preference by setting BB_SKIP_ANSI_EMULATION if desired.

By a quirk of fate, old MSYS2 programs (from 2021) which required
a workaround in busybox-w32 (commit 54d2ea4b4) are immune to the
problem with ConEmu.

(GitHib issue #287)