drop: add cdrop and pdrop aliases
Add cdrop and pdrop applets as aliases for drop.  If a command
isn't specified these use cmd.exe and PowerShell instead of the
BusyBox shell.

This makes it possible to choose the default shell used for SSH
connections even in older versions of OpenSSH that don't support
the DefaultShellArguments registry key.

Note that to get cmd.exe to run a command rather than an interactive
shell it's necessary to set the DefaultShellCommandOption registry
key to '/c'.

Costs 248-272 bytes.