ash: changes to evalsubshell()
When a spawn_forkshell() was required in evalsubshell() two calls
to expredir() were being made:  one in the parent and one in the
child.  Rearrange the code so there's only one call, in the child.

The call to expredir() in the child is necessary because copynode()
doesn't take a copy of expfname in the nfile node.  Code could be
added to do this but it's cheaper to call expredir().

Add code in forkshell_evalsubshell() to turn off the EV_TESTED
flag for background processes.  I haven't found a case where this
makes a difference but no doubt somebody would have eventually.