ash: special treatment for read builtin
When INT is being trapped the read builtin gets special treatment
in bash.  In a top-level interactive shell interrupting the read
with Ctrl-C clears the input and allows the user to enter a new
string.  In a subshell Ctrl-C really does interrupt the read and
the trap isn't executed.

zsh works similarly, except that in the latter case the trap is
executed.

dash interrupts the read and executes the trap in both cases.

ksh also interrupts the read in both cases but only executes the
trap in the first.

Implement the bash behaviour.

(GitHub issue #303)