win32: ignore ctrl-c in parent of execve(2)
The execve(2) system call is emulated for Microsoft Windows.  This
requires the creation of a new process.  The old process remains
active, waiting for the "execed" child to exit so it can pass on
its exit status.

Previously this was achieved using P_WAIT mode in the call to
spawnve().  However the parent of the execve(2) process may still
be able to catch Ctrl-C interrupts.  This can lead to unwanted
behaviour, such as a shell and its children competing for input.

Force the waiting process to ignore Ctrl-C interrupts.

Costs 64-80 bytes.

(GitHub issue #303)