xargs: kill children when interrupted by Ctrl-C
A user was running the following command in a clone of the
busybox-w32 repo from a busybox-w32 shell using Git for Windows:

   git rev-list --all | xargs git grep main.c

Using Ctrl-C to interrupt the command resulted in the familiar
problem that two processes ended up competing for user input:
in this case the busybox-w32 shell and the Git for Windows pager,
less.

Make the problem go away (in this particular case) by killing all
the children of xargs when it's interrupted.  This is probably a
sensible thing to do for other circumstances too.

Adds 64-100 bytes.

(GitHub issue #306)