ash: more Unix-style path fixes
Commit 950b318a2 (ash: Unix-style paths, shell builtins and applets)
broke running of Unix-style absolute paths from the shell by PATH
search.  The copy of the program name on the stack should only be
used in the first call to tryexec() otherwise it interferes with the
use of the stack by padvance().

Also, bring the code to handle Unix-style paths in find_command()
into line with the revised code in shellexec().