su: allow command file and arguments
The real su allows a command file to be run, with arguments given
on the command line.  Similarly it allows the script run with '-c'
to take arguments.

Permit the same in the Windows implementation.  For compatibility
this requires a user name to be specified, even though we only
support elevation.  The user name must be 'root'.

Adds 128-144 bytes.

(GitHub PR #317)