su: pass additional arguments to sh
The real su allows a command file to be run, with arguments given
on the command line.  Similarly it allows the script run with '-c'
to take arguments.  In fact, any arguments can be passed to sh.

Permit the same in the Windows implementation.  For compatibility
this requires a user name to be specified, even though we only
support elevation.  The user name must be 'root'.

Also, ensure the '-c' command is properly quoted.

Adds 144-152 bytes.

(GitHub PR #317)