win32: code shrink quote_args()
Replace parts of quote_args() with code from avih's GitHub
PR #317.

This overestimates the size of the buffer to avoid having to
calculate the exact size.

Retain the code to determine whether the argument needs to be
quoted.  Quoting arguments unconditionally wastes space on the
command line and causes the test "xargs argument line too long"
to fail.

Saves 144-176 bytes.