su: add option to keep console open on shell exit
The '-N' option keeps the console window open after the elevated
shell exits.

Previously this was achieved by passing the '-s' option along with
'-c'.  Recent changes allow scripts to be run without using '-c'
so a new mechanism is required.

su passes the '-N' flag to the shell.  This causes the shell to
issue a prompt and wait for user input when it exits.

Costs 200-204 bytes.