win32: crtl-c interrupts non-console applications
Commit 9db9b34ad (win32: ignore ctrl-c in parent of execve(2))
prevented a parent process from reacting to Ctrl-C while it was
waiting for its child to complete.  This avoids the problem where
a shell and an interactive child end up competing for input after
a Ctrl-C.

However, a child process which isn't attached to the console
(a GUI application, for example) can't then be killed by Ctrl-C.

Instead of completely ignoring Ctrl-C give the parent a handler
which detects if its child is attached to the console.  If so it's
left to handle Ctrl-C itself and the parent ignores the interrupt.
If not the parent terminates the child and all its children as if
by SIGINT.

Costs 200 bytes.