ash: enable 'set -/+o noconsole'
Previously the 'noconsole' shell option could only be set as a
shell command line option.  Allow it to be changed from within
the shell by 'set -o noconsole' or 'set +o noconsole'.

The console window is now minimised rather than hidden.  This
makes it easier for the user to access the console when 'noconsole'
is in effect.

Adds 8-32 bytes.

(GitHub issue #325)