win32: another BB_OVERRIDE_APPLETS fix
Support for conditionally replacing applets with external commands
requires the ability to check whether a given command name is
present on PATH.  This was being done using the PATH environment
variable, which works in commands run by the shell but not in the
shell itself.

- The shell uses the *shell* variable PATH to look for executables.
  This may not be the same as the *environment* variable.

- 'command -p' uses an entirely different PATH.

Applet look-up in the shell is now treated as a special case, with
the actual PATH being used passed to the look-up code in a global
variable.

This doesn't affect tab completion in the shell:  whether a
completion is an applet or an external command is irrelevant.

Costs 152-288 bytes.

(GitHub issue #329)