win32: code shrink
getsysdir() wasn't used in get_proc_addr() because the former
called realpath() which in turn called the latter.  (Commit
4c6c8d61bc)

realpath() was used to adjust the case of the path name as it
was visible to the user as root's home directory.  (Commit
b04bbc0109)

Later the home directory for "root" was changed so it no longer
needed getsysdir().  (Commit 385decd6bf)

The remaining uses of getsysdir() don't require the path to be
canonicalised, so realpath() can be removed from getsysdir() and
getsysdir() can be used in get_proc_addr().

Saves 32-64 bytes